Migrated refreshCodedTests
